+++ packages/admin/src/form/SubmitResult.tsx
@@ -1,0 +1,12 @@
+export class SubmitError extends Error {
+    submitError?: any;
+
+    constructor(message?: string, submitError?: any) {
+        super(message);
+        this.submitError = submitError;
+    }
+}
+
+export interface SubmitResult {
+    error?: SubmitError;
+}